alpha 0.0.6 b2b

+++ Models/B2BBalance.cs
@@ -6,9 +6,9 @@ namespace UFOPay.Models
         public int userId { get; set; }
         public int kassaId { get; set; }
         public bool b2bAccess { get; set; } = false;
-        public int b2b_balance_usd { get; set; }
-        public int b2b_balance_eur { get; set; }
-        public int b2b_balance_pln { get; set; }
-        public int b2b_balance_rub { get; set; }
+        public double b2b_balance_usd { get; set; }
+        public double b2b_balance_eur { get; set; }
+        public double b2b_balance_pln { get; set; }
+        public double b2b_balance_rub { get; set; }
     }
 }
